Separate scalems.radical._common.py.
Common interfaces with only rp dependencies are in _common.py.
scalems.radical.raptor may include scalems-specific details and may
depend on both radical.pilot and on arbitrary scalems package features.
